Refactor `State` types into `prefect.states` (#7108)
# Conflicts:
#	src/prefect/blocks/core.py
#	src/prefect/deployments.py
