+++ src/prefect/client/__init__.py
@@ -7,7 +7,7 @@ Explore the client by communicating with an in-memory webserver - no setup requi
 ```
 $ # start python REPL with native await functionality
 $ python -m asyncio
->>> from prefect.client import get_client
+>>> from prefect.client.orion import get_client
 >>> async with get_client() as client:
 ...     response = await client.hello()
 ...     print(response.json())
@@ -15,7 +15,3 @@ $ python -m asyncio
 ```
 </div>
 """
-from prefect.client.orion import get_client, OrionClient
-from prefect.client.cloud import get_cloud_client, CloudClient
-
-__all__ = ["get_client", "OrionClient", "get_cloud_client", "CloudClient"]
