+++ tests/test_futures.py
@@ -5,10 +5,10 @@ from unittest.mock import MagicMock
 import pytest
 
 from prefect.client import OrionClient
-from prefect.client.schemas import Completed
 from prefect.deprecated.data_documents import DataDocument
 from prefect.flows import flow
 from prefect.futures import PrefectFuture, resolve_futures_to_data
+from prefect.states import Completed
 from prefect.tasks import task
 from prefect.testing.utilities import assert_does_not_warn
