+++ tests/test_tasks.py
@@ -10,7 +10,6 @@ import pytest
 
 from prefect import flow, get_run_logger, tags
 from prefect.blocks.core import Block
-from prefect.client.schemas import State
 from prefect.context import PrefectObjectRegistry, TaskRunContext, get_run_context
 from prefect.deprecated.data_documents import DataDocument
 from prefect.engine import get_state_for_result
@@ -23,6 +22,7 @@ from prefect.futures import PrefectFuture
 from prefect.orion import models
 from prefect.orion.schemas.core import TaskRunResult
 from prefect.orion.schemas.states import StateType
+from prefect.states import State
 from prefect.tasks import Task, task, task_input_hash
 from prefect.testing.utilities import exceptions_equal, flaky_on_windows
 from prefect.utilities.annotations import allow_failure, unmapped
